requirements.txt
beautifulsoup4
python-dotenv
faiss-cpu
ipykernel
jupyter
langchain
langchain-community
langchain-core
langchain-openai
openai
sentence-transformers
streamlit
tiktoken